pyproject.toml
[project]
name = "c592-mmha-net"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.129.0",
    "google-generativeai>=0.8.6",
    "jinja2>=3.1.6",
    "numpy>=2.2.6",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.7.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "uvicorn>=0.41.0",
]
